Move all source code folders into src
