+++ src/pages/learn.js
@@ -3,14 +3,14 @@ import Image from 'next/image'
 import Link from 'next/link'
 import ReactAudioPlayer from 'react-audio-player'
 
-import frequency1 from '../images/frequency2.png'
-import orca from '../images/orca2.png'
-import organization1 from '../images/partner1.png'
-import organization2 from '../images/partner2.png'
-import roundorca from '../images/roundorca.png'
-import salishsea from '../images/salishsea.png'
-import audio from '../public/frequency.mp3'
-import orcasound from '../public/frequency.mp3'
+import frequency1 from '../../images/frequency2.png'
+import orca from '../../images/orca2.png'
+import organization1 from '../../images/partner1.png'
+import organization2 from '../../images/partner2.png'
+import roundorca from '../../images/roundorca.png'
+import salishsea from '../../images/salishsea.png'
+import audio from '../../public/frequency.mp3'
+import orcasound from '../../public/frequency.mp3'
 import learnStyles from '../styles/Learn.module.css'
 export const learn = () => {
   return (
